Create spirohash.py
Proof-of-concept of the hashing algorithm, before providing a fancy interface, help, or any other supporting documentation.

This PoC uses an approximation of the golden spiral instead of moving the pivot point at discrete positions. This makes the math more elegant at the expense of correctness. We'll see if this decision persists to the final release.